Split out losses
